src/main/java/authapi/user/domain/User.java
package authapi.user.domain;

import authapi.user.domain.value.Email;
import authapi.user.domain.value.Password;

import java.util.Objects;

public class User {
    private final Long id;
    private final String name;
    private final Email email;
    private final Password password;

    public User(Long id, String name, String email, String password) {
        this.id = id;
        this.name = name;
        this.email = new Email(email);
        this.password = new Password(password);
    }

    public Long id() {
        return id;
    }

    public String name() {
        return name;
    }

    public String email() {
        return email.value();
    }

    public String password() {
        return password.value();
    }

    public static UserBuilder builder() {
        return new UserBuilder();
    }

    public static class UserBuilder {
        private Long id;
        private String name;
        private String email;
        private String password;

        public UserBuilder id(Long id) {
            this.id = id;
            return this;
        }

        public UserBuilder name(String name) {
            this.name = name;
            return this;
        }

        public UserBuilder email(String email) {
            this.email = email;
            return this;
        }

        public UserBuilder password(String password) {
            this.password = password;
            return this;
        }

        public User build() {
            return new User(id, name, email, password);
        }
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        User user = (User) o;
        return Objects.equals(id, user.id) && name.equals(user.name) && email.equals(user.email) && password.equals(user.password);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, name, email, password);
    }
}
